Refactor remaining common components into $lib
