+++ src/projects/ProjectPin.svelte
@@ -16,7 +16,7 @@
 <script lang="ts">
   import { getCsrfToken } from "../api/session.js";
   import { baseApiUrl } from "../api/base.js";
-  import Pin from "../common/Pin.svelte";
+  import Pin from "../lib/components/common/Pin.svelte";
 
   export let project: Project;
   export let size = 1;
